+++ io/task.c
@@ -25,42 +25,39 @@
 #include "qom/object.h"
 #include "trace.h"
 
-struct QIOTaskThreadData {
-    QIOTaskWorker worker;
-    gpointer opaque;
-    GDestroyNotify destroy;
-    GMainContext *context;
-    GSource *completion;
-};
-
-
-struct QIOTask {
-    Object *source;
-    QIOTaskFunc func;
-    gpointer opaque;
-    GDestroyNotify destroy;
-    Error *err;
-    gpointer result;
-    GDestroyNotify destroyResult;
-    QemuMutex thread_lock;
-    QemuCond thread_cond;
-    struct QIOTaskThreadData *thread;
-};
-
-
-QIOTask *qio_task_new(Object *source,
-                      QIOTaskFunc func,
-                      gpointer opaque,
-                      GDestroyNotify destroy)
+struct QIOTaskThreadData
 {
-    QIOTask *task;
+    QIOTaskWorker  worker;
+    gpointer       opaque;
+    GDestroyNotify destroy;
+    GMainContext*  context;
+    GSource*       completion;
+};
+
+struct QIOTask
+{
+    Object*                   source;
+    QIOTaskFunc               func;
+    gpointer                  opaque;
+    GDestroyNotify            destroy;
+    Error*                    err;
+    gpointer                  result;
+    GDestroyNotify            destroyResult;
+    QemuMutex                 thread_lock;
+    QemuCond                  thread_cond;
+    struct QIOTaskThreadData* thread;
+};
+
+QIOTask* qio_task_new(Object* source, QIOTaskFunc func, gpointer opaque, GDestroyNotify destroy)
+{
+    QIOTask* task;
 
     task = g_new0(QIOTask, 1);
 
     task->source = source;
     object_ref(source);
-    task->func = func;
-    task->opaque = opaque;
+    task->func    = func;
+    task->opaque  = opaque;
     task->destroy = destroy;
     qemu_mutex_init(&task->thread_lock);
     qemu_cond_init(&task->thread_cond);
@@ -70,30 +67,20 @@ QIOTask *qio_task_new(Object *source,
     return task;
 }
 
-static void qio_task_free(QIOTask *task)
+static void qio_task_free(QIOTask* task)
 {
     qemu_mutex_lock(&task->thread_lock);
     if (task->thread) {
-        if (task->thread->destroy) {
-            task->thread->destroy(task->thread->opaque);
-        }
+        if (task->thread->destroy) { task->thread->destroy(task->thread->opaque); }
 
-        if (task->thread->context) {
-            g_main_context_unref(task->thread->context);
-        }
+        if (task->thread->context) { g_main_context_unref(task->thread->context); }
 
         g_free(task->thread);
     }
 
-    if (task->destroy) {
-        task->destroy(task->opaque);
-    }
-    if (task->destroyResult) {
-        task->destroyResult(task->result);
-    }
-    if (task->err) {
-        error_free(task->err);
-    }
+    if (task->destroy) { task->destroy(task->opaque); }
+    if (task->destroyResult) { task->destroyResult(task->result); }
+    if (task->err) { error_free(task->err); }
     object_unref(task->source);
 
     qemu_mutex_unlock(&task->thread_lock);
@@ -103,10 +90,9 @@ static void qio_task_free(QIOTask *task)
     g_free(task);
 }
 
-
 static gboolean qio_task_thread_result(gpointer opaque)
 {
-    QIOTask *task = opaque;
+    QIOTask* task = opaque;
 
     trace_qio_task_thread_result(task);
     qio_task_complete(task);
@@ -114,10 +100,9 @@ static gboolean qio_task_thread_result(gpointer opaque)
     return FALSE;
 }
 
-
 static gpointer qio_task_thread_worker(gpointer opaque)
 {
-    QIOTask *task = opaque;
+    QIOTask* task = opaque;
 
     trace_qio_task_thread_run(task);
 
@@ -133,10 +118,8 @@ static gpointer qio_task_thread_worker(gpointer opaque)
     qemu_mutex_lock(&task->thread_lock);
 
     task->thread->completion = g_idle_source_new();
-    g_source_set_callback(task->thread->completion,
-                          qio_task_thread_result, task, NULL);
-    g_source_attach(task->thread->completion,
-                    task->thread->context);
+    g_source_set_callback(task->thread->completion, qio_task_thread_result, task, NULL);
+    g_source_attach(task->thread->completion, task->thread->context);
     g_source_unref(task->thread->completion);
     trace_qio_task_thread_source_attach(task, task->thread->completion);
 
@@ -146,43 +129,30 @@ static gpointer qio_task_thread_worker(gpointer opaque)
     return NULL;
 }
 
-
-void qio_task_run_in_thread(QIOTask *task,
-                            QIOTaskWorker worker,
-                            gpointer opaque,
-                            GDestroyNotify destroy,
-                            GMainContext *context)
+void qio_task_run_in_thread(QIOTask* task, QIOTaskWorker worker, gpointer opaque, GDestroyNotify destroy,
+                            GMainContext* context)
 {
-    struct QIOTaskThreadData *data = g_new0(struct QIOTaskThreadData, 1);
-    QemuThread thread;
+    struct QIOTaskThreadData* data = g_new0(struct QIOTaskThreadData, 1);
+    QemuThread                thread;
 
-    if (context) {
-        g_main_context_ref(context);
-    }
+    if (context) { g_main_context_ref(context); }
 
-    data->worker = worker;
-    data->opaque = opaque;
+    data->worker  = worker;
+    data->opaque  = opaque;
     data->destroy = destroy;
     data->context = context;
 
     task->thread = data;
 
     trace_qio_task_thread_start(task, worker, opaque);
-    qemu_thread_create(&thread,
-                       "io-task-worker",
-                       qio_task_thread_worker,
-                       task,
-                       QEMU_THREAD_DETACHED);
+    qemu_thread_create(&thread, "io-task-worker", qio_task_thread_worker, task, QEMU_THREAD_DETACHED);
 }
 
-
-void qio_task_wait_thread(QIOTask *task)
+void qio_task_wait_thread(QIOTask* task)
 {
     qemu_mutex_lock(&task->thread_lock);
     assert(task->thread != NULL);
-    while (task->thread->completion == NULL) {
-        qemu_cond_wait(&task->thread_cond, &task->thread_lock);
-    }
+    while (task->thread->completion == NULL) { qemu_cond_wait(&task->thread_cond, &task->thread_lock); }
 
     trace_qio_task_thread_source_cancel(task, task->thread->completion);
     g_source_destroy(task->thread->completion);
@@ -191,24 +161,16 @@ void qio_task_wait_thread(QIOTask *task)
     qio_task_thread_result(task);
 }
 
-
-void qio_task_complete(QIOTask *task)
+void qio_task_complete(QIOTask* task)
 {
     task->func(task, task->opaque);
     trace_qio_task_complete(task);
     qio_task_free(task);
 }
 
+void qio_task_set_error(QIOTask* task, Error* err) { error_propagate(&task->err, err); }
 
-void qio_task_set_error(QIOTask *task,
-                        Error *err)
-{
-    error_propagate(&task->err, err);
-}
-
-
-bool qio_task_propagate_error(QIOTask *task,
-                              Error **errp)
+bool qio_task_propagate_error(QIOTask* task, Error** errp)
 {
     if (task->err) {
         error_propagate(errp, task->err);
@@ -219,23 +181,12 @@ bool qio_task_propagate_error(QIOTask *task,
     return false;
 }
 
-
-void qio_task_set_result_pointer(QIOTask *task,
-                                 gpointer result,
-                                 GDestroyNotify destroy)
+void qio_task_set_result_pointer(QIOTask* task, gpointer result, GDestroyNotify destroy)
 {
-    task->result = result;
+    task->result        = result;
     task->destroyResult = destroy;
 }
 
+gpointer qio_task_get_result_pointer(QIOTask* task) { return task->result; }
 
-gpointer qio_task_get_result_pointer(QIOTask *task)
-{
-    return task->result;
-}
-
-
-Object *qio_task_get_source(QIOTask *task)
-{
-    return task->source;
-}
+Object* qio_task_get_source(QIOTask* task) { return task->source; }
